Add: Change return type of Selector::select to std::set<File*>
Changed the Selector::select from std::vector<File*> to std::set<File*>.
This change will facilitate File* manipulation in the server.
The server has two inner loop, one for the readers and one for the
writers.
The readers loop can provoke a disconnect of one of the client,
if the socket is released in the readers and isn't removed from the
vector, it can provoke a segfault.
Because of that, the use of a std::set<File*> is a better choice, the
removing is O(log(N)) compared to vector where a shift is O(n).